+++ angular.json
@@ -4,9 +4,15 @@
   "newProjectRoot": "projects",
   "projects": {
     "ngx-nestable": {
+      "projectType": "application",
+      "schematics": {
+        "@schematics/angular:component": {
+          "style": "scss"
+        }
+      },
       "root": "",
       "sourceRoot": "src",
-      "projectType": "application",
+      "prefix": "ngx",
       "architect": {
         "build": {
           "builder": "@angular-devkit/build-angular:browser",
@@ -40,6 +46,18 @@
                 {
                   "replace": "src/environments/environment.ts",
                   "with": "src/environments/environment.prod.ts"
+                }
+              ],
+              "budgets": [
+                {
+                  "type": "initial",
+                  "maximumWarning": "2mb",
+                  "maximumError": "5mb"
+                },
+                {
+                  "type": "anyComponentStyle",
+                  "maximumWarning": "6kb",
+                  "maximumError": "10kb"
                 }
               ]
             }
@@ -84,11 +102,24 @@
           "options": {
             "tsConfig": [
               "src/tsconfig.app.json",
-              "src/tsconfig.spec.json"
+              "src/tsconfig.spec.json",
+              "e2e/tsconfig.json"
             ],
             "exclude": [
               "**/node_modules/**"
             ]
+          }
+        },
+        "e2e": {
+          "builder": "@angular-devkit/build-angular:protractor",
+          "options": {
+            "protractorConfig": "e2e/protractor.conf.js",
+            "devServerTarget": "ngx-nestable:serve"
+          },
+          "configurations": {
+            "production": {
+              "devServerTarget": "ngx-nestable:serve:production"
+            }
           }
         }
       }
@@ -117,7 +148,47 @@
           }
         }
       }
-    }
+    },
+      "nestable": {
+          "projectType": "library",
+          "root": "projects/nestable",
+          "sourceRoot": "projects/nestable/src",
+          "prefix": "ngx",
+          "schematics": {
+              "@schematics/angular:component": {
+                  "styleext": "scss"
+              }
+          },
+          "architect": {
+              "build": {
+                  "builder": "@angular-devkit/build-ng-packagr:build",
+                  "options": {
+                      "tsConfig": "projects/nestable/tsconfig.lib.json",
+                      "project": "projects/nestable/ng-package.json"
+                  }
+              },
+              "test": {
+                  "builder": "@angular-devkit/build-angular:karma",
+                  "options": {
+                      "main": "projects/nestable/src/test.ts",
+                      "tsConfig": "projects/nestable/tsconfig.spec.json",
+                      "karmaConfig": "projects/nestable/karma.conf.js"
+                  }
+              },
+              "lint": {
+                  "builder": "@angular-devkit/build-angular:tslint",
+                  "options": {
+                      "tsConfig": [
+                          "projects/nestable/tsconfig.lib.json",
+                          "projects/nestable/tsconfig.spec.json"
+                      ],
+                      "exclude": [
+                          "**/node_modules/**"
+                      ]
+                  }
+              }
+          }
+  }
   },
   "defaultProject": "ngx-nestable",
   "schematics": {
