--- lib/index.ts
@@ -1,2 +1,0 @@
-export { NestableComponent } from './src/nestable.component';
-export { NestableModule } from './src/nestable.module';
