--- lib/scss-bundle.config.json
@@ -1,4 +1,0 @@
-{
-  "entry": "./lib/src/_nestable-theme.scss",
-  "dest": "./dist/ngx-nestable/_nestable-theme.scss"
-}
