--- lib/src/nestable-drag-handle/nestable-drag-handle.directive.spec.ts
@@ -1,20 +1,0 @@
-import { ElementRef } from '@angular/core';
-import {TestBed, ComponentFixture, inject} from '@angular/core/testing';
-
-import { NestableDragHandleDirective } from './nestable-drag-handle.directive';
-
-describe('NestableDragHandleDirective', () => {
-
-  const elementRef = {} as ElementRef;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      declarations: [ElementRef, NestableDragHandleDirective]
-    });
-  });
-
-  it('should create an instance', () => {
-    const directive = new NestableDragHandleDirective(elementRef);
-    expect(directive).toBeTruthy();
-  });
-});
