--- lib/src/nestable-expand-collapse-handle/nestable-expand-collapse.directive.spec.ts
@@ -1,8 +1,0 @@
-import { NestableExpandCollapseDirective } from './nestable-expand-collapse.directive';
-
-describe('NestableExpandCollapseDirective', () => {
-  it('should create an instance', () => {
-    const directive = new NestableExpandCollapseDirective();
-    expect(directive).toBeTruthy();
-  });
-});
