--- lib/src/nestable.component.scss
@@ -1,58 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=Roboto');
-
-ul {
-    .dd-placeholder { margin: 5px 0; padding: 0; min-height: 30px; background: #f2fbff; border: 1px dashed #b6bcbf; box-sizing: border-box;}
-
-    li {
-
-        .nestable-item-mask {
-            display: none;
-            position: absolute;
-            top: 0; bottom: 0;
-            right: 0; left: 0;
-            z-index: 9998;
-        }        
-
-        .nestable-expand-button { 
-            display: block; 
-            position: relative;            
-            cursor: pointer; 
-            float: left; 
-            width: 25px; 
-            height: 14px; 
-            padding: 0; 
-            white-space: nowrap; 
-            overflow: hidden; 
-            border: 0; 
-            background: transparent; 
-            font-size: 18px; 
-            line-height: 1; 
-            text-align: center; 
-            font-weight: bold; 
-            outline: 0;
-        }
-
-        .nestable-item-container {
-            position: relative;
-            display: flex;
-            flex-direction: row;
-            align-items: center;
-            color: rgba(0,0,0,.87);
-            min-height: 32px;
-            font-size: 16px;
-            font-family: 'Roboto', sans-serif;
-            cursor: pointer;
-            outline: 0;
-            margin-bottom: 2px;
-            padding-left: 8px;
-        }
-
-        .nestable-item-container:hover {
-            background: rgba(0,0,0,.04);
-        }
-    }
-}
-
-ol, ul {
-	list-style: none;
-}
