+++ package.json
@@ -23,56 +23,56 @@
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "pack": "rimraf dist && ng-packagr -p ng-package.json && npm run bundle-styles && npm run readme-to-dist",
-    "bundle-styles": "scss-bundle -c lib/scss-bundle.config.json",
+    "pack": "rimraf dist && ng build nestable && npm run bundle-styles && npm run readme-to-dist",
+    "bundle-styles": "scss-bundle -c projects/nestable/scss-bundle.config.json",
     "readme-to-dist": "copyfiles README.md dist/ngx-nestable",
     "publish-package": "npm run pack && cd dist/ngx-nestable/ && ls && npm publish"
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "4.4.6",
-    "@angular/cdk": "2.0.0-beta.12",
-    "@angular/common": "4.4.6",
-    "@angular/compiler": "4.4.6",
-    "@angular/core": "4.4.6",
-    "@angular/forms": "4.4.6",
-    "@angular/http": "4.4.6",
-    "@angular/material": "2.0.0-beta.12",
-    "@angular/platform-browser": "4.4.6",
-    "@angular/platform-browser-dynamic": "4.4.6",
-    "core-js": "2.5.1",
-    "rxjs": "5.5.2",
-    "zone.js": "0.8.18",
-    "@angular/flex-layout": "2.0.0-beta.10"
+    "@angular/animations": "~8.2.0",
+    "@angular/cdk": "^8.1.2",
+    "@angular/common": "~8.2.0",
+    "@angular/compiler": "~8.2.0",
+    "@angular/core": "~8.2.0",
+    "@angular/flex-layout": "^8.0.0-beta.26",
+    "@angular/forms": "~8.2.0",
+    "@angular/material": "^8.1.3",
+    "@angular/platform-browser": "~8.2.0",
+    "@angular/platform-browser-dynamic": "~8.2.0",
+    "@angular/router": "~8.2.0",
+    "rxjs": "~6.4.0",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "~1.2.6",
-    "@angular/compiler-cli": "4.4.6",
-    "@angular/language-service": "^4.4.6",
-    "@ngtools/webpack": "1.5.6",
-    "@types/jasmine": "2.5.38",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "6.0.92",
-    "codelyzer": "~2.0.0",
-    "copyfiles": "^2.0.0",
-    "jasmine-core": "~2.5.2",
-    "jasmine-spec-reporter": "~3.2.0",
-    "karma": "~1.4.1",
-    "karma-chrome-launcher": "~2.0.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "0.2.3",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "0.2.2",
+    "@angular-devkit/build-angular": "~0.802.1",
+    "@angular-devkit/build-ng-packagr": "~0.802.1",
+    "@angular/cli": "~8.2.1",
+    "@angular/compiler-cli": "~8.2.0",
+    "@angular/language-service": "~8.2.0",
+    "@types/node": "~8.9.4",
+    "@ngtools/webpack": "8.2.1",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "codelyzer": "^5.0.0",
+    "copyfiles": "^2.1.1",
+    "jasmine-core": "~3.4.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.1.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-cli": "~2.0.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "live-server": "1.2.0",
-    "ng-packagr": "1.6.0",
-    "protractor": "~5.1.0",
-    "rimraf": "^2.6.2",
-    "scss-bundle": "^2.1.3",
-    "sw-precache": "5.2.0",
-    "ts-node": "~2.0.0",
-    "tsickle": "^0.35.0",
-    "tslint": "~4.5.0",
-    "typescript": "2.3.4"
+    "ng-packagr": "^5.3.0",
+    "protractor": "~5.4.0",
+    "rimraf": "^3.0.0",
+    "scss-bundle": "^2.5.1",
+    "ts-node": "~7.0.0",
+    "tsickle": "^0.36.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.5.3"
   }
 }
