+++ projects/nestable/src/lib/nestable.module.ts
@@ -5,6 +5,7 @@ import { NestableComponent } from './nestable.component';
 import { NestableDragHandleDirective } from './nestable-drag-handle/nestable-drag-handle.directive';
 
 import { NestableExpandCollapseDirective } from './nestable-expand-collapse-handle/nestable-expand-collapse.directive';
+
 @NgModule({
     imports: [CommonModule],
     declarations: [
@@ -18,4 +19,5 @@ import { NestableExpandCollapseDirective } from './nestable-expand-collapse-hand
         NestableExpandCollapseDirective
     ]
 })
-export class NestableModule { }
+export class NestableModule {
+}
