+++ projects/nestable/src/public-api.ts
@@ -1,0 +1,6 @@
+/*
+ * Public API Surface of nestable
+ */
+
+export * from './lib/nestable.component';
+export * from './lib/nestable.module';
